83 - Simple Refactoring for slightly improved modularity
