+++ ui/src/components/Login.vue
@@ -117,7 +117,7 @@ export default {
     submit: function (event) {
       this.isLoading = true
       this.errorMessage = this.successMessage = ''
-      this.$http.post(`${this.$globals.api}/login`, {email: this.email, password: this.password}).then(function (res) {
+      this.$http.post(`${this.$globals.api}/auth/login`, {email: this.email, password: this.password}).then(function (res) {
         if (res.body.messageType === 'loginSuccess') {
           this.successMessage = res.body.payload.message
           this.$globals.startSession(res.body.payload)
